+++ src/pages/pessoa/pessoa.list.page.tsx
@@ -7,6 +7,7 @@ import CustomTable, {
 } from "../../components/custom-table/custom-table.component";
 import EnumTipoPessoa from "../../models/enums/enumTipoPessoa";
 import Pessoa from "../../models/entities/pessoa";
+import AlertMessage from "../../helpers/alertMessages";
 
 const columns: Column[] = [
 	{ id: "id", label: "Id" },
@@ -56,11 +57,17 @@ export default function PessoaListPage() {
 	};
 
 	const handleRemove = (id: string) => {
-		PessoaService.remove(id).then((response) => {
-			if (response.success) {
-				if (paginationData?.items.length === 1 && page > 0)
-					setPage(page - 1);
-				else getData(page, rowsPerPage);
+		AlertMessage.confirmRemoveItem().then((response) => {
+			if (response.isConfirmed) {
+				PessoaService.remove(id).then((response) => {
+					if (response.success) {
+						if (paginationData?.items.length === 1 && page > 0)
+							setPage(page - 1);
+						else getData(page, rowsPerPage);
+
+						AlertMessage.showRemoveSuccess();
+					}
+				});
 			}
 		});
 	};
@@ -70,11 +77,15 @@ export default function PessoaListPage() {
 	};
 
 	function getData(page: number, rowsPerPage: number) {
-		PessoaService.getAll(page + 1, rowsPerPage).then((response) => {
-			if (response.success) {
-				setPaginationData(response.data);
+		PessoaService.getAll(page + 1, rowsPerPage).then(
+			(response) => {
+				if (response.success) setPaginationData(response.data);
+				else AlertMessage.showError(response.errorMessage);
+			},
+			(error) => {
+				AlertMessage.showError(error);
 			}
-		});
+		);
 	}
 
 	useEffect(() => {
